requirements.txt
asgiref==3.6.0
astroid==2.15.5
colorama==0.4.6
crispy-bootstrap5==0.7
dill==0.3.6
Django==4.2.1
django-bootstrap5==23.1
django-crispy-forms==2.0
django-extensions==3.2.3
django-fontawesome-5==1.0.18
django-tinymce==3.6.1
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.1.1
phonenumbers==8.13.11
Pillow==9.5.0
platformdirs==3.5.1
pydot==1.4.2
pylint==2.17.4
pyparsing==3.0.9
pypng==0.20220715.0
qrcode==7.4.2
regex==2023.5.5
sqlparse==0.4.4
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata==2023.3
wrapt==1.15.0
